[ft 162198528] add 84 of test coverage locally

+++ src/controller/Parcel-controller.js
@@ -8,14 +8,13 @@ const Parcels = {
   // Create a parcel delivery order
   async create(req, res) {
     const {
-      location, destination, presentLocation, length, width, height,
+      location, destination, presentLocation, weight,
     } = req.body;
-
-    const newParcel = new Parcel(uuidv4(), location, destination, presentLocation, length, width,
-      height, req.user.id, 'Pending', moment(new Date()), moment(new Date()));
+    const newParcel = new Parcel(uuidv4(), location, destination, presentLocation, weight,
+      req.user.id, 'Pending', moment(new Date()), moment(new Date()));
     const createQuery = `INSERT INTO
-      parcels(id, location, destination ,present_location, length, width, height, owner_id, status, created_date, modified_date)
-      VALUES($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11)
+      parcels(id, location, destination ,present_location, weight, owner_id, status, created_date, modified_date)
+      VALUES($1, $2, $3, $4, $5, $6, $7, $8, $9)
       returning *`;
     try {
       const { rows, rowCount } = await db.query(createQuery, Object.values(newParcel));
@@ -30,11 +29,11 @@ const Parcels = {
   },
   // Fetch all parcel delivery orders
   async getAll(req, res) {
-    const findAllQuery = 'SELECT * FROM parcels';
+    const findAllQuery = 'SELECT * FROM parcels WHERE owner_id = $1';
     try {
-      const { rows, rowCount } = await db.query(findAllQuery);
+      const { rows, rowCount } = await db.query(findAllQuery, [req.user.id]);
       return res.status(200).send({
-        message: 'Success', status: 200, rowCount, data: rows[0],
+        message: 'Success', status: 200, rowCount, data: rows,
       });
     } catch (error) {
       return res.status(400).send({
@@ -44,9 +43,9 @@ const Parcels = {
   },
   // Fetch all parcel delivery orders by a specific user
   async parcelByUser(req, res) {
-    const parcelByUserQuery = 'SELECT * FROM parcels where owner_id = $1';
+    const parcelByUserQuery = 'SELECT * FROM parcels WHERE owner_id = $1';
     try {
-      const { rows, rowCount } = await db.query(parcelByUserQuery, [req.params.userId]);
+      const { rows, rowCount } = await db.query(parcelByUserQuery, [req.user.id]);
       return res.status(200).send({
         message: 'Success', status: 200, rowCount, data: rows[0],
       });
@@ -58,9 +57,9 @@ const Parcels = {
   },
   // Fetch a specific parcel delivery order
   async getOne(req, res) {
-    const text = 'SELECT * FROM parcels WHERE id = $1';
+    const text = 'SELECT * FROM parcels WHERE id = $1 AND owner_id = $2';
     try {
-      const { rows, rowCount } = await db.query(text, [req.params.parcelId]);
+      const { rows, rowCount } = await db.query(text, [req.params.parcelId, req.user.id]);
       if (!rows[0]) {
         return res.status(404).send({ message: 'parcels not found', status: 404 });
       }
@@ -75,12 +74,12 @@ const Parcels = {
   },
   // Cancel the specific parcel delivery order
   async cancel(req, res) {
-    const findOneQuery = 'SELECT * FROM parcels WHERE id = $1';
+    const findOneQuery = 'SELECT * FROM parcels WHERE id = $1 AND owner_id = $2';
     const updateOneQuery = `UPDATE parcels
       SET status=$1,modified_date=$2
-      WHERE id=$3 returning *`;
+      WHERE id=$3 AND owner_id = $4 returning *`;
     try {
-      const { rows } = await db.query(findOneQuery, [req.params.parcelId]);
+      const { rows } = await db.query(findOneQuery, [req.params.parcelId, req.user.id]);
       if (!rows[0]) {
         return res.status(404).send({ message: 'parcel not found' });
       }
@@ -88,6 +87,7 @@ const Parcels = {
         'Cancelled',
         moment(new Date()),
         rows[0].id,
+        req.user.id,
       ];
       const response = await db.query(updateOneQuery, updateValues);
       return res.status(200).send({
@@ -99,12 +99,12 @@ const Parcels = {
   },
   // Change the present location of a specific parcel delivery order
   async ChangePresentLocation(req, res) {
-    const findOneQuery = 'SELECT * FROM parcels WHERE id = $1';
+    const findOneQuery = 'SELECT * FROM parcels WHERE id = $1 AND owner_id = $2';
     const updateOneQuery = `UPDATE parcels
       SET present_location=$1,modified_date=$2
-      WHERE id=$3 returning *`;
+      WHERE id=$3 AND owner_id = $4 returning *`;
     try {
-      const { rows } = await db.query(findOneQuery, [req.params.parcelId]);
+      const { rows } = await db.query(findOneQuery, [req.params.parcelId, req.user.id]);
       if (!rows[0]) {
         return res.status(404).send({ message: 'Parcel not found' });
       }
@@ -112,6 +112,7 @@ const Parcels = {
         req.body.present_location,
         moment(new Date()),
         rows[0].id,
+        req.user.id,
       ];
       const response = await db.query(updateOneQuery, updateValues);
       return res.status(200).send({
@@ -121,15 +122,13 @@ const Parcels = {
       return res.status(400).send(err);
     }
   },
-  // Change the location ofa specific parcel delivery order -
-  // only for the user who created it
   async changeDestination(req, res) {
-    const findOneQuery = 'SELECT * FROM parcels WHERE id = $1';
+    const findOneQuery = 'SELECT * FROM parcels WHERE id = $1 AND owner_id = $2';
     const updateOneQuery = `UPDATE parcels
       SET destination=$1,modified_date=$2
-      WHERE id=$3 returning *`;
+      WHERE id=$3 AND owner_id = $4 returning *`;
     try {
-      const { rows } = await db.query(findOneQuery, [req.params.parcelId]);
+      const { rows } = await db.query(findOneQuery, [req.params.parcelId, req.user.id]);
       if (!rows[0]) {
         return res.status(404).send({ message: 'Parcel not found' });
       }
@@ -137,22 +136,24 @@ const Parcels = {
         req.body.destination,
         moment(new Date()),
         rows[0].id,
+        req.user.id,
       ];
       const response = await db.query(updateOneQuery, updateValues);
       return res.status(200).send({
         message: 'Success', status: 200, data: response.rows[0],
       });
     } catch (err) {
+      console.log(err);
       return res.status(400).send(err);
     }
   },
   async changeStatus(req, res) {
-    const findOneQuery = 'SELECT * FROM parcels WHERE id = $1';
+    const findOneQuery = 'SELECT * FROM parcels WHERE id = $1 AND owner_id = $2';
     const updateOneQuery = `UPDATE parcels
       SET status=$1,modified_date=$2
-      WHERE id=$3 returning *`;
+      WHERE id=$3 AND owner_id = $4 returning *`;
     try {
-      const { rows } = await db.query(findOneQuery, [req.params.parcelId]);
+      const { rows } = await db.query(findOneQuery, [req.params.parcelId, req.user.id]);
       if (!rows[0]) {
         return res.status(404).send({ message: 'Parcel not found' });
       }
@@ -160,6 +161,7 @@ const Parcels = {
         req.body.status,
         moment(new Date()),
         rows[0].id,
+        req.user.id,
       ];
       const response = await db.query(updateOneQuery, updateValues);
       return res.status(200).send({
@@ -170,9 +172,9 @@ const Parcels = {
     }
   },
   async delete(req, res) {
-    const deleteQuery = 'DELETE FROM parcels WHERE id = $1 returning *';
+    const deleteQuery = 'DELETE FROM parcels WHERE id = $1 AND owner_id = $2 returning *';
     try {
-      const { rows } = await db.query(deleteQuery, [req.params.parcelId]);
+      const { rows } = await db.query(deleteQuery, [req.params.parcelId, req.user.id]);
       if (!rows[0]) {
         return res.status(404).send({ message: 'parcels not found', status: 404 });
       }
